refactor(environment-groups): move to react
