+++ app/edge/EdgeDevices/EdgeDevicesView/EdgeDevicesDatatable/EdgeDevicesDatatable.tsx
@@ -4,7 +4,7 @@ import _ from 'lodash';
 
 import { Environment } from '@/react/portainer/environments/types';
 import { AMTDevicesDatatable } from '@/edge/EdgeDevices/EdgeDevicesView/AMTDevicesDatatable/AMTDevicesDatatable';
-import { EnvironmentGroup } from '@/portainer/environment-groups/types';
+import { EnvironmentGroup } from '@/react/portainer/environments/environment-groups/types';
 
 import { PaginationControls } from '@@/PaginationControls';
 import {
