+++ app/react/portainer/environments/environment-groups/environment-groups.service.ts
@@ -1,4 +1,4 @@
-import axios, { parseAxiosError } from '../services/axios';
+import axios, { parseAxiosError } from '@/portainer/services/axios';
 
 import { EnvironmentGroup, EnvironmentGroupId } from './types';
